Runs in Python 3
Based on Rezende and Mohamed (2016). Variational Inference with Normalizing Flows is an approach that makes it possible to do VI with a set of very flexible, non-convex and multimodal approximate distributions. This is an implementation of the algorithm with an application to a toy problem.

The objective is to obtain optimal values for five of the nine variational parameters in the model by applying stochastic gradient ascent.

Rezende, D., J., Mohamed, S. (June 14, 2016). Variational Inference with Normalizing Flows. Google DeepMind London. Retrieved from https://arxiv.org/pdf/1505.05770.pdf.