Runs in Python 3
I focus on a country with 10 cities and a delivery company that wants to send packages from one city to another more efficiently. I implemented a simulation in which cities are represented as nodes and roads as edges in a graph. Each node and each edge have an attribute that contains a list of elements of class packages. Each package has the following characteristics: origin, destination, and current position. Roads are two-way and each way has its own maximum capacity of transporting packages. Each city also has a maximum capacity for storing packages. I am measuring the total number of steps it takes to deliver all packages under different conditions.

One assumption that I make is that the probability of one package to want to go to any other city is the same. In real life, most packages will want to arrive at larger cities - those with larger capacities. Similarly, I draw the road capacities from a uniform distribution, and in real life, there is a larger probability that roads between large cities have larger capacities. The maximum capacity of a city, however, is dependent on its connectivity - the more cities it is directly connected to, proportionally the more storage capacity for packages it has. The maximum capacity of each city is drawn from a binomial distribution, centered at a proportion of a theoretical limit. This proportion depends on a saturation parameter and the aforementioned rule.

The parameters of my simulation are the maximum road capacity, the maximum distance between two cities, and the maximum city storage capacity. The output is the number of steps it takes to deliver all packages.