Runs in Python 2
I worked on the classic Wumpus World game that we studied in class. I developed an AI that can play the game (and often win!), and a simple interface, which a human can use to play as well. I am using the power of the resolution algorithm in propositional logic as a sound and complete algorithm for logical inferences.

The agent walks around the Wumpus World and collects percepts. These are two types - cell-specific (’breeze’, ’stench’, ’glitter’), or non-cell-specific, or action-specific (’bump’, ’scream’). It makes logical inferences based on these percepts on whether each cell certainly has, or certainly does not have a threat, such as a pit or a Wumpus.