Runs in Python 2
The Range Minimum Query (RMQ) problem asks for the minimum value in a sublist of a given list of size n.

Yuan & Atallah (2010) present an O(n) solution to the RMQ problem that can be generalized to solve the AMQ problem in O(mn) time. It only works for lists of size a power of two, but it can be modified to work for lists of any size. Also, it only optimizes for the number of comparisons, not running time. For example, the query process makes at most one comparison but takes O(logn) time to run because of the recursion. The preprocessing procedure makes O(n) comparisons but takes O(nlogn) time to run.

Yuan, H., & Atallah, M. J. (2010). Data Structures for Range Minimum Queries in Multidimensional Arrays. Proceedings of the Twenty-First Annual ACM-SIAM Symposium on Discrete Algorithms.